Initial refactor to change example file names.
