+++ runelite-plugin.properties
@@ -1,6 +1,6 @@
-displayName=Example
-author=Nobody
+displayName=SkillLevelNotifications
+author=JC Arbelbide (jcarbelbide on github)
 support=
-description=An example greeter plugin
-tags=
-plugins=com.example.ExamplePlugin
+description=Notifies client when skills fall below a threshold.
+tags=skill,notifier,notifications,mage,magic,reduced,reduction,level
+plugins=com.skilllevelnotifications.SkillLevelNotificationsPlugin
