--- src/test/java/com/example/ExamplePluginTest.java
@@ -1,13 +1,0 @@
-package com.example;
-
-import net.runelite.client.RuneLite;
-import net.runelite.client.externalplugins.ExternalPluginManager;
-
-public class ExamplePluginTest
-{
-	public static void main(String[] args) throws Exception
-	{
-		ExternalPluginManager.loadBuiltin(ExamplePlugin.class);
-		RuneLite.main(args);
-	}
-}
